requirements.txt
ultralytics
roboflow
opencv-python
matplotlib
pyyaml
